# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Agent OS API
  version: 0.6.0
  description: >-
    Control plane and observability API for the Agent OS SPOG. Route
    surface is generated authoritatively from the chi router; payload
    shapes are defined by the SQL migrations (internal/db/migrations)
    and the sqlc query layer (internal/db/queries).
servers:
  - url: http://localhost:8080
    description: Local / hpms1 deployment
...
paths:
...
  /api/agents/{id}:
    delete:
      tags: [agents]
      operationId: delete_DeleteAgent
      summary: Delete Agent
      description: Handler DeleteAgent in internal/api.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Success }
        '400': { description: Bad request }
        '404': { description: Not found }
        '500': { description: Server error }
    get:
      tags: [agents]
      operationId: get_GetAgent
      summary: Get Agent
      description: Handler GetAgent in internal/api.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Success }
        '400': { description: Bad request }
        '404': { description: Not found }
        '500': { description: Server error }
    patch:
      tags: [agents]
      operationId: patch_UpdateAgentConfig
      summary: Update Agent Config
      description: Handler UpdateAgentConfig in internal/api.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Success }
        '400': { description: Bad request }
        '404': { description: Not found }
        '500': { description: Server error }
...